fix: dark mode issue and api endpoint for ecosystem
