+++ src/app/SideBar.astro
@@ -822,7 +822,7 @@ import { pathRoutes } from '../config/pathRoutes';
 		</div>
 
 		<div
-			class="fixed inset-0 z-10 hidden bg-gray-900/50 dark:bg-gray-900/90"
+			class="fixed inset-0 -z-10 hidden bg-gray-900/50 dark:bg-gray-900/90"
 			id="sidebarBackdrop"
 		>
 		</div>
